+++ call.py
@@ -1,11 +1,24 @@
-from getch import getch #py-getch
 from datetime import datetime
 from time import sleep
-import subprocess
 import tempfile
 import sys, os
 import glob
 import json
+try:
+    from getch import getch #py-getch
+    from colorama import init, Fore, Back, Style
+except ModuleNotFoundError:
+    print("You are missing dependencies do you want to install them with pip? [Y/n]")
+    cmd = input("#")
+    if(cmd == "y" or cmd == "Y"):
+        os.system("pip install -r requirements.txt")
+        print("\n\nif all went well, restart yt-dl and it will work\nIf not run \"pip install -r requirements.txt\"\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+    else:
+        print("install them before you can use yt-dl\nYou should also install ffmpeg and add it to PATH")
+        input("press any key to quit")
+
+init()
 
 def is_venv():
     return (hasattr(sys, 'real_prefix') or
@@ -13,14 +26,14 @@ def is_venv():
 
 year = datetime.now().year
 curb = "master"
-ver = f"2.1.4-{curb}"
-lstupdt = "2020-04-25"
+ver = f"2.1.5-{curb}" #lang(2python3) #featureset #patch/bugfix pre, RC
+lstupdt = "2020-05-30"
 spath = sys.path[0]+os.path.sep
 settings = spath+"settings.json"
 
 #==========NAME==========#
-def name():
-    print(f"yt-dl {ver} by KoleckOLP (C){year}\n")
+def name(newline=True):
+    BC("yt-dl {ver} by KoleckOLP (C){year}\n", newline)
 
 #==========MUSTYPLAT CLEAR==========#
 def clear():
@@ -43,6 +56,8 @@ def readchar(o): #multiplatform readchar
 #==========FIRST RUN MENU==========#
 def firstrun():
     clear()
+    print("This program requires one more dependenci that we can't install trough pip, ffmpeg you need to get this one your self and add it to path")
+    input("Press any key if you are aware.")
     print("What's the name of your python executable.\n<enter> for python")
     py = input("#")
     if (py == ""):
@@ -62,58 +77,70 @@ def firstrun():
     if (cmd == "y"):
         aup = True
     else:
-        pass
+        aup = False
     print()
     savepath("chp",py,pip,ydpip,aup)
     loadpath()
-    print("Do you want a Launch script? [Y/n]")
+    print("Do you want a Launch script? [Y/n] or p=" + Fore.BLUE + "Powershell" + Style.RESET_ALL)
     cmd = readchar("#")
     if (cmd == "y"):
         launchs()
+    elif(cmd== "p"):
+        launchs(True)
     else:
         pass
     loadpath("hid")
     about()
 
 #==========MAKE LAUNCH SCRIPT==========#
-def launchs():
+def launchs(p=""):
     print(f"Do you have venv set up if yes type name of the venv")
     cmd = input("#")
-    if(cmd != ""):
-        if(os.name == 'nt'):
-            f=open("yt-dl.bat","w")
-            f.write(f"@echo off\n\ncmd /c \"cd /d {spath}{cmd}{os.path.sep}Scripts & activate & cd /d {spath} & {py} main.py\"")
-            f.close()
-        elif(os.name == 'posix'):
-            f=open("yt-dl.sh","w")
-            f.write(f"#!/bin/bash\n\ncd {spath}{cmd}{os.path.sep}bin && source activate && cd {spath} && {py} main.py")
+    if(p==True):
+        if(cmd != ""):
+            f=open("yt-dl.ps1","w")
+            f.write(f"Set-Location {spath}{cmd}{os.path.sep}Scripts\n.{os.path.sep}Activate.ps1\nSet-Location {spath}\n{py} main.py")
             f.close()
         else:
-            print('####If you see this please contact the dev. 0x1015####')
+            f=open("yt-dl.bat","w")
+            f.write(f"Set-Location {spath}\n{py} main.py")
+            f.close()
     else:
-        if(os.name == 'nt'):
-            f=open("yt-dl.bat","w")
-            f.write(f"@echo off\n\ncmd /c \"cd /d {spath} & {py} main.py\"")
-            f.close()
-        elif(os.name == 'posix'):
-            f=open("yt-dl.sh","w")
-            f.write(f"#!/bin/bash\n\ncd {spath} && {py} main.py")
-            f.close()
+        if(cmd != ""):
+            if(os.name == 'nt'):
+                f=open("yt-dl.bat","w")
+                f.write(f"@echo off\n\ncmd /c \"cd /d {spath}{cmd}{os.path.sep}Scripts & activate & cd /d {spath} & {py} main.py\"")
+                f.close()
+            elif(os.name == 'posix'):
+                f=open("yt-dl","w")
+                f.write(f"#!/bin/bash\n\ncd {spath}{cmd}{os.path.sep}bin && source activate && cd {spath} && {py} main.py")
+                f.close()
+            else:
+                print('####If you see this please contact the dev. 0x1015####')
         else:
-            print('####If you see this please contact the dev. 0x1005####')
+            if(os.name == 'nt'):
+                f=open("yt-dl.bat","w")
+                f.write(f"@echo off\n\ncmd /c \"cd /d {spath} & {py} main.py\"")
+                f.close()
+            elif(os.name == 'posix'):
+                f=open("yt-dl","w")
+                f.write(f"#!/bin/bash\n\ncd {spath} && {py} main.py")
+                f.close()
+            else:
+                print('####If you see this please contact the dev. 0x1005####')
 
 #==========ABOUT==========#
 def about():
     clear()
-    print(f"yt-dl version: {ver} by KoleckOLP,\n"
-         +f"HorseArmored inc (C){year}\n"
+    name(False)
+    print(f"HorseArmored inc (C){year}\n"
          +f"Last updated on: {lstupdt}\n"
          +f"My webpage: https://koleckolp.comli.com/\n"
          +f"Project page: https://github.com/KoleckOLP/yt-dl\n"
          +f"youtube-dl (C)2008-2011 Ricardo Garcia Gonzalez\n"
          +f"           (C)2011-{year} youtube-dl developers\n"
          +f"ffmpeg (C)2000-{year} FFmpeg team")
-    print("Do you want to see whats new? [Y/n]")
+    print(Style.BRIGHT + "Do you want to see whats new? [Y/n]" + Style.RESET_ALL)
     cmd = readchar("#")
     if (cmd == "y"): 
         clear()
@@ -153,7 +180,7 @@ def savepath(a="chp", x="", y="", z="", q=""):
                     json.dump({"audio": aud+os.path.sep,"videos": vid+os.path.sep,"py": x,"pip": y,"ydpip": z,"aup": q}, fh)
                     fh.close()
     if (a != "chp"):
-        loadpath()
+        loadpath("hid")
         fh = open(settings, "w")
         json.dump({"audio": audio,"videos": videos,"py": x,"pip": y,"ydpip": z,"aup": q}, fh)
         fh.close()
@@ -191,13 +218,15 @@ def loadpath(s="show"):
         ydpip = path["ydpip"]
         aup = path["aup"]
     if(s == "show"):
-        print(f"audio is saved to: {audio}\nvideo is saved to: {videos}\n")
+        print(Style.BRIGHT + "audio is saved to: " + Style.RESET_ALL, end="")
+        print(audio)
+        print(Style.BRIGHT + "video is saved to: " + Style.RESET_ALL, end="")
+        print(videos + "\n")
 
 #==========SAVE MENU==========#
 def slpath():
     loadpath()
-    global aup
-    print(f"1. change download path\n2. delete settings\n3. change autoupdate={aup}\n4. generate Launch script\n0. GoBack")
+    print("1. change download path\n2. delete settings\n3. generate Launch script\n0. GoBack")
     cmd = readchar("#")
     if (cmd == "1"):
         savepath()
@@ -207,10 +236,14 @@ def slpath():
         os.remove(settings)
         firstrun()
     elif (cmd == "3"):
-        aup = not aup
-        savepath("",py,pip,ydpip,aup)
-    elif (cmd == "4"):
-        launchs()
+        print("press p=" + Fore.BLUE + "PowerShell" + Style.RESET_ALL + " or <Enter>")
+        cmd=input("#")
+        if(cmd=="p"):
+            launchs(True)
+        else:
+            launchs()
+        clear()
+        name()
     else:
         clear()
         name()
@@ -227,42 +260,58 @@ def upytdl():
 
 #==========UPDATE DEPEND==========#
 def upyd():
+    print("updating pip...")
+    os.system(f"{py} -m {pip} install --upgrade {pip}")
     print("Updating dependencies...")
-    os.system(f"{pip} install -r requirements.txt")
+    os.system(f"{pip} install -U -r requirements.txt")
 
 #==========UPDATE MENU==========#
 def update():
     clear()
+    global aup
     if(ydpip == True):
-        print("What do you want to update?\n1. All\n2. yt-dl\n3. dependencies\n4. change branch\n0. GoBack")
+        print(Style.BRIGHT + f"What do you want to update?" + Style.RESET_ALL + "\n1. All\n2. yt-dl\n3. dependencies\n4. change autoupdate=", end="")
+        TF(aup, False)
+        print(f"\n5. change branch=", end="")
+        BC(curb)
+        print("\n0. GoBack")
         cmd = readchar("#")
         if(cmd == "1"): #All
             clear()
             upytdl()
             upyd()
-            print("\n!!!restart for chanes to take effect!!!\n")
+            print(Fore.RED, "\n!!!restart for changes to take effect!!!\n", Style.RESET_ALL)
         elif(cmd == "2"): #yt-dl
             clear()
             upytdl()
-            print("\n!!!restart for chanes to take effect!!!\n")
+            print(Fore.RED, "\n!!!restart for changes to take effect!!!\n", Style.RESET_ALL)
         elif(cmd == "3"): #youtube-dl
             clear()
             upyd()
             print("")
-        elif(cmd == '4'):
+        elif (cmd == "4"):
+            aup = not aup
+            savepath("",py,pip,ydpip,aup)
+            loadpath("hid")
+            clear()
+            print(f"autoupdate={aup}\n")
+        elif(cmd == '5'):
             clear()
             print(f"What branch do you wat to change to? You are in a {curb}")
             if (curb == "master"):
                 otherb = "testing"
             else:
                 otherb = "master"
-            print(f"do you want to switch to {otherb} [Y/n]")
+            print(f"do you want to switch to ")
+            BC(otherb, False, True)
+            print(" [Y/n]")
             cmd = readchar("#")
             if (cmd == "y"):
                 os.system(f"git checkout {otherb}")
-                print("\n!!!restart for chanes to take effect!!!\n")
+                print(Fore.RED, "\n!!!restart for changes to take effect!!!\n", Style.RESET_ALL)
             else:
-                pass
+                clear()
+                name()
         else:
             clear()
             name()
@@ -271,7 +320,8 @@ def update():
 
 #==========AUTO UPDATE==========#
 def autoupdt():
-    print(f"autoupdate={aup}")
+    print(f"autoupdate=", end="")
+    TF(aup)
     if (aup == True):
         upytdl()
         upyd()
@@ -374,35 +424,102 @@ def subd():
         os.system(f"ffmpeg -i \"{pie[0]}\" \"{lick}\"")
         print("\a")
 
+#==========VID TO HEVC==========#
+def vidhevc():
+    clear()
+    print(Fore.RED + "Files with special charactes in the path may not work, also keep filenames short" + Style.RESET_ALL)
+    print("<Enter> single video 1. numbered series or 0. GoBack")
+    cmd = input("#")
+    if(cmd == ""): #single
+        print("write path to the file you want to reencode")
+        url = input("#")
+        print("reenceded file will get \"_lib265.mkv\" appended, or type a different one")
+        append = input("#")
+        if(append == ""):
+            append = "_lib265.mkv"
+        os.system(f"ffmpeg -hwaccel auto -i \"{url}\" -map 0:v -map 0:a? -map 0:s? -c:v libx265 -rc constqp -qp 24 -b:v 0K -c:a aac -b:a 384k -c:s copy \"{os.path.splitext(url)[0]+append}\"")
+        print("\a")
+    elif(cmd == "1"): #numbered
+        print("write path to the folder with videos don't forget to add \\*.extencion")
+        url = input("#")
+        print("type a short path of the episode name just before the number")
+        common_name = input("#")
+        print("Write the highest number of the video")
+        numb_last = input("#")
+        try:
+            numb_last = int(numb_last)
+        except ValueError:
+            clear()
+            print("That's not a number")
+            name()
+        print("Write the number you want to start from <Enter> for 1 or 0. GoBack")
+        numb = input("#")
+        if(numb == ""):
+            numb = 1
+        else:
+            try:
+                numb = int(numb)
+            except ValueError:
+                clear()
+                print("That's not a number")
+                name()
+        print("does the numbers use zero padding [Y/n]")
+        zero = readchar("#")
+        print("\nreenceded file will get \"_lib265\" appended, or type a different one")
+        append = input("#")
+        if(append == ""):
+            append = "_lib265"
+        url = url.replace('[', '[[]')
+        episodes = glob.glob(url)
+        print(episodes)
+        for numb in range(numb, numb_last+1):
+            if(zero == "y"):
+                if(numb < 10):
+                        numb_final = "0"+str(numb)
+                        numb_final = common_name+str(numb_final)
+                else:
+                    numb_final = numb
+                    numb_final = common_name+str(numb_final)
+            for i in episodes:
+                if str(numb_final) in i:
+                    filename = os.path.basename(i)
+                    file_split = filename.split(".", 1)
+                    path = os.path.dirname(i)
+                    finali = path+os.path.sep+file_split[0]+append+".mkv"
+                    os.system(f"ffmpeg -hwaccel auto -i \"{i}\" -map 0:v -map 0:a? -map 0:s? -c:v libx265 -rc constqp -qp 24 -b:v 0K -c:a aac -b:a 384k -c:s copy \"{finali}\"")
+            print("\a")
+    else:
+        clear()
+        name()
+
 #==========DEBUG CONSOLE==========#
 def debug():
-    print("Welcome to debug menu:")
+    print(Fore.MAGENTA + "Welcome to debug menu:", Style.RESET_ALL)
     while(True):
         cmd = input(">")
-        if (cmd == "paths"):
+        if(cmd == "all"):
             print("Paths:")
             for p in sys.path: print(p)
-        elif(cmd == "saves"):
-            print("Saves:")
+            print("\nSaves:")
             if(os.path.exists(os.path.realpath(os.path.dirname(sys.argv[0]))+os.path.sep+".git")):
                 git = True
             else:
                 git = False
-            print(f"audio is saved to: {audio}\nvideo is saved to: {videos}\npython executable name: {py}\npip executable name: {pip}\nyoutube-dl from pip: {ydpip}\nyt-dl from git: {git}\nautoupdate: {aup}")
-        elif(cmd == "all"):
-            print("Paths:")
-            for p in sys.path: print(p)
-            print("Saves:")
-            if(os.path.exists(os.path.realpath(os.path.dirname(sys.argv[0]))+os.path.sep+".git")):
-                git = True
-            else:
-                git = False
-            print(f"audio is saved to: {audio}\nvideo is saved to: {videos}\npython executable name: {py}\npip executable name: {pip}\nyoutube-dl from pip: {ydpip}\nyt-dl from git: {git}\nautoupdate: {aup}")
+            loadpath()
+            print("Variables:")
+            print(f"python executable name: {py}\npip executable name: {pip}")
+            print("youtube-dl from pip: ", end="")
+            TF(ydpip, False)
+            print("\nyt-dl from git: ", end="")
+            TF(git, False)
+            print("\nautoupdate: ", end="")
+            TF(aup)
             if is_venv():
                 venv = True
             else:
                 venv = False
-            print(f"in venv: {venv}")
+            print("in venv: ", end="")
+            TF(venv)
         elif(cmd == "deldown"):
             print("Are you sure you want to delete all audio and videos [Y/n]")
             cmd = readchar("")
@@ -418,4 +535,37 @@ def debug():
         else:
             clear()
             name()
-            break    
+            break
+
+#==========COLOR FUNCTIONS==========#
+def TF(var="", newline=True):
+    if newline == True:
+        end = "\n"
+    else:
+        end = ""
+  
+    if (var == True):
+        print(Fore.GREEN + str(var) + Style.RESET_ALL, end=end)
+    elif (var == False):
+        print(Fore.RED + str(var) + Style.RESET_ALL, end=end)
+    else:
+        print("Not a boolean", end=end)
+
+def BC(stri="no input", newline=True, reverse=False):
+    if newline == True:
+        end = "\n"
+    else:
+        end = ""
+
+    if(reverse == True):
+        if(curb == "master"):
+            branch = "testing"
+        elif(curb == "testing"):
+            branch = "master"
+    else:
+        branch = curb  
+
+    if(branch == "master"):
+        print(Fore.CYAN + eval(f'f"""{stri}"""') + Style.RESET_ALL, end=end)
+    else:
+        print(Fore.MAGENTA + eval(f'f"""{stri}"""') + Style.RESET_ALL, end=end)